% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GetServerArgs}
\alias{GetServerArgs}
\title{Initialise shared memory on client and obtain server launch string.}
\usage{
GetServerArgs(matrixDimension, withVectors, numGPUsWanted, memName, semName,
  printDetails)
}
\arguments{
\item{matrixDimension}{- type (integer) - the dimension of the (assumed square) matrix}

\item{withVectors}{- type (boolean) - true = We want the eigenvectors and eigenvalues calculated ; false = we only want eigenvalues calculated}

\item{numGPUsWanted}{- type (string)  - The number of GPUs to use in for the symmetric eigenvalue (syevd) computation}

\item{memName}{- type (string)  - a name to give to the named shared memory region (will be created in /dev/shm/) and defaults to the user name if nothing specified}

\item{semName}{- type (string)  - a name to give to the semaphore (will be placed in /dev/shm) and defaults to the user name if nothing specified}

\item{printDetails}{- type (integer 0|1|2) - 0 = don't print, 1 = print details of server progress to screen; 2 = print to log (not functional)}
}
\value{
- type (string) A string that can be used a command line arguments to run the syevd_server executable
}
\description{
This function will create a CSharedMemory object that initialises the shared memory region and the semaphore used for comms beteween client and server.
If the object is already initialised it is removed and reinitialised.
Returns a string of the form "-n 10000 -v 1 -g 3 -m /syevx_<PID_of_client> -s /sem_<PID_of_client> -p" that can be used to launch a syevd_server process
that will accept matrix data on which to perform eigenvalue decomposition
}

