% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/MagmaEigenNonsym.R
\name{MakeServer}
\alias{MakeServer}
\title{Creates the server executable.}
\usage{
MakeServer(environmentSetup = "", target = "all")
}
\arguments{
\item{environmentSetup}{- type (string) e.g. "env LD_LIBRARY_PATH=/usr/local/magma-1.7.1:$LD_LIBRARY_PATH "}

\item{target}{- type (string) The make target e.g. all | clean | dist-clean | install}
}
\value{
A character vector containing output of the make process
}
\description{
Creates syevd_server executable using a call to 'make' and the makefile and make.inc information present in the <package root>/src directory.
Users have to set the following variable in make.inc :
     MAGMALIB  =$(MAGMA_HOME)/lib          # The path to the MAGMA library

Users must ensure that MAGMA_HOME and possibly CUDA_ROOT environment variables have been set in the shells environment
or the variables can be set using environmentSetup paramater.
e.g. "environmentSetup="env MAGMA_HOME=/data/bow355/A_275ERCP_R_MAGMA/magma-1.7.0 CUDA_ROOT=/cm/shared/apps/cuda65/toolkit/6.5.14 "
}

